+++ ChangeLog.md
@@ -13,7 +13,7 @@
 - hpacucli: parse controller status, 03ae764
 - hpacucli: check/report controller/cache/battery status, f16f0c2
 - fix plugin commandline options support, a2c5b8a
-- hpacucli: handle not configured controller with noraid status code. [#145], [#151], 85423dd
+- hpacucli: handle not configured controller with noraid status code. [#145], [#151], [#154], 85423dd
 
 ## [4.0.2] - 2016-03-14